+++ src/utils/getColorOptions.ts
@@ -1,9 +1,9 @@
 import * as vscode from "vscode";
 import { getIconPath } from "./getIconPath";
-import colors from "../lists/colors.json";
+import { getContributedColors } from "./getContributedColors";
 
 export const getColorOptions = (context: vscode.ExtensionContext) => {
-  return colors.map(({ id, description, defaults }) => ({
+  return getContributedColors(context).map(({ id, description, defaults }) => ({
     label: description,
     description: id,
     iconPath: getIconPath(context, defaults.dark),